alot

+++ reorder.py
@@ -13,7 +13,22 @@ import pyperclip
 def get_parent_dir(directory):
     return os.path.dirname(directory)
 
-def reorder_csv(filename_in, filename_out):
+def reorder_csv_3D(filename_in, filename_out):
+    with open(filename_in, 'r') as infile, open(filename_out, 'a') as outfile:
+
+        # output dict needs a list for new column ordering
+        fieldnames = ['field.header.stamp', 'field.point.x', 'field.point.y',
+                      'field.point.z',
+                      '%time', 'field.header.seq', 'field.header.frame_id']
+        writer = csv.DictWriter(outfile, fieldnames=fieldnames)
+
+        # reorder the header first
+        writer.writeheader()
+        for row in csv.DictReader(infile):
+            # writes the reordered rows to the new file
+            writer.writerow(row)
+
+def reorder_csv_6D(filename_in, filename_out):
     with open(filename_in, 'r') as infile, open(filename_out, 'a') as outfile:
 
         # output dict needs a list for new column ordering
@@ -35,26 +50,31 @@ def reorder_csv(filename_in, filename_out):
 # set configuration
 # (place bag in a separate folder)
 
-bag = "/home/rm/Documents/master_thesis/data/sim2/2017-05-02-22-30-28.bag"
-topic = "/firefly/vi_sensor/ground_truth/transform" #ground truth
+bag = "/home/rm/Documents/master_thesis/data/blender/exp5_1/exp5.bag"
+gt = "sim_blender" # set to leica (3D) resp. to vicon or sim or sim_blender (6D)
+bag_leica = "/home/rm/Documents/master_thesis/data/leica_outdoor/31_ST_L_2017-05-05-11-11-38.bag"
+topic = "/leica/position" #ground truth
+
+#in case of sim
 img_topic = "/firefly/vi_sensor/left/image_raw"
 imu_topic = "/firefly/vi_sensor/imu"
-config = "/home/rm/catkin_ws_jonathan/src/mvisplanner/interface_mvisp/config/parameter/okvis_visensor_simulation.yaml"
+
+config = "/home/rm/catkin_ws/src/okvis_ros/okvis/config/config_blender.yaml"
 egm_config = "/home/rm/Documents/master_thesis/src/tools/egm.yaml"
 protocol_template = "/home/rm/Documents/master_thesis/src/tools/protocol.txt"
 
 #######################################################
-
-dirname = os.path.dirname(bag)
-command = "rosrun rosbag topic_renamer.py " + img_topic + " " + bag + " /cam0/image_raw " + dirname + "/temp.bag"
-command += " && rosrun rosbag topic_renamer.py " + imu_topic + " "+ dirname + "/temp.bag" + " /imu0 " + dirname + "/sim.bag"
-command += " && rm " + dirname + "/temp.bag && rm " + bag
-bag = dirname + "/sim.bag"
-print(command)
-pyperclip.copy(command)
-print("paste in terminal")
-raw_input('Press Enter')
-print
+if gt == "sim":
+    dirname = os.path.dirname(bag)
+    command = "rosrun rosbag topic_renamer.py " + img_topic + " " + bag + " /cam0/image_raw " + dirname + "/temp.bag"
+    command += " && rosrun rosbag topic_renamer.py " + imu_topic + " "+ dirname + "/temp.bag" + " /imu0 " + dirname + "/sim.bag"
+    command += " && rm " + dirname + "/temp.bag && rm " + bag
+    bag = dirname + "/sim.bag"
+    print(command)
+    pyperclip.copy(command)
+    print("paste in terminal")
+    raw_input('Press Enter')
+    print
 
 okvis_output = "/home/rm/Documents/master_thesis/data/okvis_output"
 if not os.path.exists(okvis_output):
@@ -65,29 +85,44 @@ else:
 
 directory = get_parent_dir(bag)
 
+# ground truth bag to csv
+if gt == "vicon" or gt == "sim":
+    csv_file = directory + "/vicon_data.csv"
 
-# bag to csv
-csv_file = directory + "/vicon_data.csv"
+    command = "rostopic echo " + "-b " + bag + " -p " +  topic + " >" + csv_file
+    print(command)
+    pyperclip.copy(command)
+    print("paste in terminal")
+    raw_input('Press Enter')
+    print
 
-command = "rostopic echo " + "-b " + bag + " -p " +  topic + " >" + csv_file
-print(command)
-pyperclip.copy(command)
-print("paste in terminal")
-raw_input('Press Enter')
-print
+    # reorder csv column
+    csv_file_reordered = csv_file[:-4] + '_reordered.csv'
+    reorder_csv_6D(csv_file, csv_file_reordered)
 
-#pBagToCsv = subprocess.Popen( ["rostopic", "echo",
-#                               "-b", bag,
-#                               "-p",  topic,
-#                               ">", csv_file,
-#                               ])
-#pBagToCsv.wait()
+elif gt == "leica":
+    csv_file = directory + "/leica_data.csv"
 
-# reorder csv column
-csv_file_reordered = csv_file[:-4] + '_reordered.csv'
-reorder_csv(csv_file, csv_file_reordered)
+    command = "rostopic echo " + "-b " + bag_leica + " -p " + topic + " >" + csv_file
+    print(command)
+    pyperclip.copy(command)
+    print("paste in terminal")
+    raw_input('Press Enter')
+    print
 
-# copy donfig file to bag folder
+    # reorder csv column
+    csv_file_reordered = csv_file[:-4] + '_reordered.csv'
+    reorder_csv_3D(csv_file, csv_file_reordered)
+elif gt =='sim_blender':
+    print("provide ground truth file!")
+    raw_input('Press Enter')
+
+    csv_file = directory + "/poses.csv"
+    # reorder csv column
+    csv_file_reordered = csv_file[:-4] + '_reordered.csv'
+    reorder_csv_6D(csv_file, csv_file_reordered)
+
+# copy config file to bag folder
 shutil.copy(config, directory)
 
 #run okvis synchronous
